requirements.txt
alembic==1.7.7
aniso8601==9.0.1
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.6.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
gunicorn